// src/types.ts
export interface Job {
    createdAt: string;
    jobTitle: string;
    companyName: string;
    industry: string;
    location: string;
    jobType: string;
    minExp: number;
    maxExp: number;
    minSalary: number;
    maxSalary: number;
    totalEmp: string;
    applyType: string;
    id: string;
}